validation-service/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
backoff==2.2.1
bytewax==0.21.1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
Deprecated==1.2.15
fastapi==0.115.5
googleapis-common-protos==1.66.0
grpcio==1.68.0
h11==0.14.0
idna==3.10
importlib-metadata==6.11.0
numpy==2.1.3
opentelemetry-api==1.22.0
opentelemetry-exporter-otlp==1.22.0
opentelemetry-exporter-otlp-proto-common==1.22.0
opentelemetry-exporter-otlp-proto-grpc==1.22.0
opentelemetry-exporter-otlp-proto-http==1.22.0
opentelemetry-instrumentation==0.43b0
opentelemetry-instrumentation-asgi==0.43b0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-instrumentation-requests==0.43b0
opentelemetry-proto==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-semantic-conventions==0.43b0
opentelemetry-util-http==0.43b0
packaging==24.2
pandas==2.2.3
prometheus-fastapi-instrumentator==7.0.0
prometheus_client==0.21.0
protobuf==4.25.5
pubsubplus-opentelemetry-integration==1.0.1
pydantic==2.10.2
pydantic_core==2.27.1
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
six==1.16.0
sniffio==1.3.1
solace-pubsubplus==1.9.0
starlette==0.41.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.1
wrapt==1.17.0
zipp==3.21.0
